opensc-explorer: warning 'expects type "unsigned int"'

+++ src/tools/opensc-explorer.c
@@ -1409,7 +1409,7 @@ static int do_asn1(int argc, char **argv)
 		goto err;
 	}
 	if ((size_t)r != len) {
-		printf("expecting %u, got only %d bytes.\n", len, r);
+		printf("expecting %u, got only %d bytes.\n", (unsigned int)len, r);
 		goto err;
 	}
